Minor modifs for win-x64 compatibility
git-svn-id: svn+ssh://svn.code.sf.net/p/tango-cs/code/share/yat/trunk@17940 4e9c00fd-8f2e-0410-aa12-93ce3db5e235

+++ src/system/Win32SysUtilsImpl.cpp
@@ -155,11 +155,11 @@ bool SysUtils::exec(const char* pszCmdLine, const char *pszDefDir, int bBackgrou
   if( !bBackground && bIsExecutable)
   {
     // Waiting for the process is done
-    uint32 uiReturnCode;
+    DWORD uiReturnCode;
     ::WaitForSingleObject( aProcessInformation.hProcess, INFINITE);
     ::GetExitCodeProcess(aProcessInformation.hProcess, &uiReturnCode);
     if( puiReturnCode )
-      *puiReturnCode = uiReturnCode;
+      *puiReturnCode = static_cast<int>(uiReturnCode);
     return true;
   }
